move tests out of internet tests
